% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{clear}
\alias{clear}
\title{Clears all items from the local storage.}
\usage{
clear(session = getDefaultReactiveDomain())
}
\arguments{
\item{session}{The Shiny session object}
}
\description{
Clears all items from the local storage.
}
\examples{
if(interactive()) {
 observeEvent(input$clear, {
  clear()
 })
}
}
