% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{configureShinyLocalStorage}
\alias{configureShinyLocalStorage}
\title{Configures the ShinyLocalStorage instance for a given page / app.}
\usage{
configureShinyLocalStorage(appName, session = getDefaultReactiveDomain())
}
\arguments{
\item{appName}{A unique identifier for the app. This is used to ensure that
the ShinyLocalStorage instance is unique to this app.}

\item{session}{The Shiny session object}
}
\description{
This must be called from within the server function of a Shiny app.
You are not required to configure ShinyLocalStorage, but it is recommended
to get a unique data store for each app.
}
\examples{
if (interactive()) {
  library(shiny)

  shinyApp(
    ui = fluidPage(
      useShinyLocalStorage(),  # Set up ShinyLocalStorage
      actionButton("btn", "Click me"),
      textInput("element", "Store me locally")
    ),
    server = function(input, output) {
      configureShinyLocalStorage("my_app")
      observeEvent(input$btn, {
        # Store the value into the local storage
        store("stored_element", input$element)
      })
    }
  )
}
}
