% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{remove}
\alias{remove}
\title{Removes a value from local storage.}
\usage{
remove(key, session = getDefaultReactiveDomain())
}
\arguments{
\item{key}{The key of the value to remove}

\item{session}{The Shiny session object}
}
\description{
Note: the namespaced \code{input} object will still exist.
}
\examples{
if(interactive()) {
 observeEvent(input$remove, {
   remove("stored_element") 
 })
} 
}
