% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{retrieve}
\alias{retrieve}
\title{Retrieves an item from local storage.}
\usage{
retrieve(key, session = getDefaultReactiveDomain())
}
\arguments{
\item{key}{The key to store the value under}

\item{session}{The Shiny session object}

\item{value}{The value to store}
}
\value{
The value stored under the key
}
\description{
Stores the retrieved value into the Shiny \code{input} object
named using a random prefix and the key, as well as returns it.
}
\details{
If \code{configureShinyLocalStorage} has not been called,
the value will be stored in the \code{input} object under the key.
}
\examples{
if(interactive()) {
 observeEvent(input$btn, {
    # Store the value into the local storage
    stored_value <- retrieve("stored_element")
 })
}
}
