% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyLS.R
\name{setLocalStorage}
\alias{setLocalStorage}
\title{setLocalStorage}
\usage{
setLocalStorage(session, ..., .dots)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer}

\item{.dots}{A list of values you want to add to localStorage in the user's browser.}

\item{\dots}{Any values you want to add to localStorage in the user's browser.}
}
\description{
Set localStorage values in the user's browser what will be accessible by the
app the next time the user accesses the page.
}
\details{
Each value passed to \code{setLocalStorage} should be named. Any unnamed value
will be silently dropped. Values set with \code{setLocalStorage} can be accessed
in the server from \code{input$localStorage}. For example, if you set \code{x = 5}
with \code{setLocalStorage} you can access that value from \code{input$localStorage$x}.
}

