% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{store}
\alias{store}
\title{Stores an item into the local storage.}
\usage{
store(key, value, session = getDefaultReactiveDomain())
}
\arguments{
\item{key}{The key to store the value under}

\item{value}{The value to store}

\item{session}{The Shiny session object}
}
\description{
Stores an item into the local storage.
}
\examples{
if(interactive()) {
 observeEvent(input$btn, {
    # Store the value into the local storage
    store("stored_element", input$element)
 })
}
}
