% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyLS.R
\name{storeInput}
\alias{storeInput}
\title{storeInput}
\usage{
storeInput(tag)
}
\arguments{
\item{tag}{An input element whose value you want to store in localStorage}
}
\description{
Automatically store and retrieve values for the wrapped input. (Experimental)
}
\details{
This function is not yet working.
}

