% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyLS.R
\name{useLocalStorage}
\alias{useLocalStorage}
\title{useLocalStorage}
\usage{
useLocalStorage()
}
\description{
Adds dependencies to shiny app necessary to access local storage from the
server function.
}
\details{
This function should be included somewhere in your shiny app's ui. It will
attach the dependencies needed to make \code{shinyLocalStorage} work. If you
use any other \code{shinyLocalStorage} functions in your ui, such as
\code{\link{storeInput}}, this function isn't needed.
}

