% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{useShinyLocalStorage}
\alias{useShinyLocalStorage}
\title{Set up a Shiny app to use ShinyLocalStorage}
\usage{
useShinyLocalStorage(container = NULL)
}
\arguments{
\item{container}{The HTML container to attach the local storage scripts to. Defaults to a div with id = "shiny-local-storage"}
}
\description{
This must be called from within a Shiny app's UI.
}
\details{
The ShinyLocalStorage Javascript library and dependencies
will be inserted as script tags in the page.
}
\examples{
if (interactive()) {
  library(shiny)

  shinyApp(
    ui = fluidPage(
      useShinyLocalStorage(),  # Set up ShinyLocalStorage
      actionButton("btn", "Click me"),
      textInput("element", "Store me locally")
    ),
    server = function(input, output) {
      observeEvent(input$btn, {
        # Store the value into the local storage
        store("stored_element", input$element)
      })
    }
  )
}
}
